+++ apps/web/src/app/layout.tsx
@@ -1,4 +1,3 @@
-import { getDatabaseAsync } from "@/db";
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
@@ -18,19 +17,11 @@ export const metadata: Metadata = {
   description: "Generated by create next app",
 };
 
-export default async function RootLayout({
+export default function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const database = await getDatabaseAsync();
-
-  const locations = await database.query.locations.findMany({});
-
-  for (const location of locations) {
-    console.log(location);
-  }
-
   return (
     <html lang="en">
       <body
