+++ apps/web/src/components/loading.tsx
@@ -1,0 +1,7 @@
+export function Loading() {
+  return (
+    <div role="status" aria-live="polite">
+      <span>Loading...</span>
+    </div>
+  );
+}
