+++ apps/web/src/db/index.ts
@@ -1,7 +1,7 @@
 import { getCloudflareContext } from "@opennextjs/cloudflare";
 import { drizzle } from "drizzle-orm/d1";
 import { cache } from "react";
-import * as schema from "./schema";
+import * as schema from "./schema/schema";
 
 export const getDatabase = cache(() => {
   const { env } = getCloudflareContext();
